Report all tags, and weekend-ness, to influxdb
The grafana stack graph doesn't like it when some values are missing
from some days - it copies the previous day.  So every tag needs a
value on every day.  But this messes up other things, as some days
I've done zero work!  So also report whether a day is a weekend day or
not.